requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
jinja2==3.1.5
python-multipart==0.0.19
python-dotenv==1.0.1
httpx==0.28.1
anthropic==0.43.0
telethon==1.37.0
aiofiles==24.1.0

# Dev / Testing
pytest==8.3.4
pytest-asyncio==0.25.0
